requirements.txt
matplotlib==3.5.0
Pillow==9.0.0
PyOpenGL
PyOpenGL_accelerate
PyYAML>=6.0
yacs>=0.1.8
scikit-image==0.19.1
termcolor
tqdm==4.62.3
loguru==0.5.3
trimesh==3.9.35
flatten_dict==0.4.2
ipykernel==5.3.4
ipywidgets==7.6.5
jpeg4py
shapely==1.7.1
vedo==2020.4.2
rtree==0.9.7
pytorch_lightning==1.2.5
PyMCubes
kornia
chumpy
cython==0.29.20
rembg>=2.0.3
opencv-python
opencv_contrib_python
git+https://github.com/Project-Splinter/human_det
git+https://github.com/YuliangXiu/smplx.git
git+https://github.com/facebookresearch/pytorch3d.git
git+https://github.com/YuliangXiu/neural_voxelization_layer.git
git+https://github.com/NVIDIAGameWorks/kaolin.git
